2024 09 05 ~ 20 BoardPost, Member 수정, BoardFile, Comment 추가 등 전체적인 수정 및  BackEnd 단 완료

+++ src/main/java/CRUD/demo/BoardPost/BoardPostForm.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class BoardPostForm {
 
-    private int BoardSequence; // sequence
+    private long BoardSequence; // sequence
     private String category;
     private String title;
     private String author;
